fix(#2444): remove incorrect sibling-function count limit
Summary
-------
Defining 256+ sibling (parallel) functions, classes, lambdas, or
comprehensions in a single scope wrongly raised "too many nested
functions in one scope (max 256)". CPython imposes no such limit, and
the cap was never about nesting depth — each function scope gets its own
`fn_protos` pool, so the count being checked was the number of protos
defined directly in one scope, not how deeply functions nest.

The 256 ceiling existed because the proto index was encoded as a `u8`
in the `MakeFunction` / `MakeClass` / `MakeClassMeta` instructions.

What changed
------------
- Widen the proto-index field from `u8` to `u32` (matching `Reg`) in the
  three Make* `Insn` variants, the `exec_make_*` VM handlers, the
  `build_def_proto` / `build_class_proto` return types, and the inliner's
  proto-index plumbing in `optimizer.rs`.
- Raise the per-scope proto limit from 256 to `u32::MAX` and reword the
  error to "too many functions in one scope (max 4294967295)" (still a
  pure defensive guard against index-space overflow; unreachable in
  practice).

Self-review caught
------------------
- The cap also applied to class bodies, lambdas, and comprehensions (all
  push into `fn_protos`), not just `def`. All four sites are fixed and
  covered by the fixture (module defs, function-body defs, class methods).
- Confirmed no proto-index truncation at the boundary: the `>= u32::MAX`
  guard fires before `len as u32` could equal `u32::MAX`.
- `MakeFunction` is a VM-dispatch op; widening one `u8` to `u32` does not
  grow the `Insn` enum (other variants dominate its size) and it is not a
  tight-loop primitive, so no hot-path regression.

Test plan
---------
- New parity fixture
  crates/pyrust/tests/cases/language/test_many_sibling_functions.py:
  300 module-scope defs, 300 defs in one function body, 300 methods in
  one class body — byte-identical to python3.12.
- cargo test --release --bin pyrust: 439 passed.
- cargo test --release --test parity_compare: green.

+++ crates/pyrust/src/optimizer.rs
@@ -639,7 +639,7 @@ fn pass_inline(
     // provided `r` is not rewritten between that point and `before_pc`.  Used to
     // identify the proto a temporary held at the moment it was copied into a
     // stable holder.
-    let make_fn_idx = |reg: u32, before_pc: usize| -> Option<u8> {
+    let make_fn_idx = |reg: u32, before_pc: usize| -> Option<u32> {
         let mut latest: Option<usize> = None;
         for (pc, insn) in insns.iter().enumerate().take(before_pc) {
             let mut w: HashSet<u32> = HashSet::new();
@@ -663,7 +663,7 @@ fn pass_inline(
     // exactly once across the whole scope (so its value is immutable for every
     // call), and that single write must either *be* a `MakeFunction` or copy in
     // a register that held one at that point.
-    let resolve_fn_reg = |reg: u32| -> Option<u8> {
+    let resolve_fn_reg = |reg: u32| -> Option<u32> {
         if write_count.get(&reg).copied() != Some(1) {
             return None;
         }
@@ -681,7 +681,7 @@ fn pass_inline(
 
     // Pre-build an inlining plan for every proto that resolves to a call target.
     // (Computed lazily / cached as we encounter call sites.)
-    let mut plans: HashMap<u8, Option<InlinePlan>> = HashMap::new();
+    let mut plans: HashMap<u32, Option<InlinePlan>> = HashMap::new();
     let mut next_window = *num_regs;
 
     // First splice pass.  `out` accumulates `(old_i, insn)` pairs where `old_i`
@@ -766,8 +766,8 @@ fn resolve_call_target(
     insns: &[Insn],
     call_pc: usize,
     fn_reg: u32,
-    resolve_fn_reg: &impl Fn(u32) -> Option<u8>,
-) -> Option<u8> {
+    resolve_fn_reg: &impl Fn(u32) -> Option<u32>,
+) -> Option<u32> {
     // Walk backwards from the call; stop at a basic-block boundary (any control
     // flow) — the call-site `Move(fn_reg, src)` is always in the same block.
     let mut i = call_pc;
